work towards picc generating specs

+++ src/block.h
@@ -25,17 +25,17 @@ class Block {
   // Will advance time on |state| by |delta| and set all register values to
   // unknown. The |range_| of this Block will be empty and it will start at the
   // advanced end time of the range supplied in |state|.
-  explicit Block(State* state, uint32 delta);
+  explicit Block(State* state, uint64 delta);
 
   // Returns 0 if this |spec| can be scheduled before this Block. Returns a time
   // within the block if this spec should be appended to this block. Returns a
   // time greater than the end of the block if the block suggests creation of a
   // new block at that time. Returns kInfinity on error.
-  uint32 EarliestTimeAfter(const Spec& spec, uint32 end_time) const;
+  uint64 EarliestTimeAfter(const Spec& spec, uint64 end_time) const;
 
   // Returns the number of clock cycles that would be consumed by appending
   // |spec| to the end of this Block.
-  uint32 ClocksToAppend(const Spec& spec) const;
+  uint64 ClocksToAppend(const Spec& spec) const;
 
   void Append(const Spec& spec);
   void AppendBlock(std::unique_ptr<Block> block);
@@ -45,14 +45,14 @@ class Block {
   const State* final_state() const { return (*states_.rbegin()).get(); };
   const Range& range() const { return range_; }
   uint32 bytes() const { return total_bytes_; }
-  uint32 clocks() const { return range_.Duration(); }
+  uint64 clocks() const { return range_.Duration(); }
 
  protected:
   Range range_;
   uint32 total_bytes_;
   std::vector<std::unique_ptr<State>> states_;
   std::vector<std::unique_ptr<op::OpCode>> opcodes_;
-  uint32 register_usage_times_[Register::REGISTER_COUNT];
+  uint64 register_usage_times_[Register::REGISTER_COUNT];
 };
 
 }  // namespace vcsmc
